Instantiate default ScoutErrorHandling implementation in Agent::fromConfig
